requirements.txt
fpdf==1.7.2
lxml==5.3.0
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
PyMuPDF==1.25.1
PyPDF2==3.0.1
python-docx==1.1.2
typing_extensions==4.12.2
